+++ src/main/java/KUSITMS/WITHUS/domain/application/position/service/PositionServiceImpl.java
@@ -4,8 +4,8 @@ import KUSITMS.WITHUS.domain.application.position.dto.PositionRequestDTO;
 import KUSITMS.WITHUS.domain.application.position.dto.PositionResponseDTO;
 import KUSITMS.WITHUS.domain.application.position.entity.Position;
 import KUSITMS.WITHUS.domain.application.position.repository.PositionRepository;
-import KUSITMS.WITHUS.domain.organization.entity.Organization;
-import KUSITMS.WITHUS.domain.organization.repository.OrganizationRepository;
+import KUSITMS.WITHUS.domain.organization.organization.entity.Organization;
+import KUSITMS.WITHUS.domain.organization.organization.repository.OrganizationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
