+++ src/main/java/KUSITMS/WITHUS/domain/organization/organization/dto/OrganizationRequestDTO.java
@@ -1,4 +1,4 @@
-package KUSITMS.WITHUS.domain.organization.dto;
+package KUSITMS.WITHUS.domain.organization.organization.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
