+++ src/main/java/KUSITMS/WITHUS/domain/organization/organization/repository/OrganizationJpaRepository.java
@@ -1,6 +1,6 @@
-package KUSITMS.WITHUS.domain.organization.repository;
+package KUSITMS.WITHUS.domain.organization.organization.repository;
 
-import KUSITMS.WITHUS.domain.organization.entity.Organization;
+import KUSITMS.WITHUS.domain.organization.organization.entity.Organization;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrganizationJpaRepository extends JpaRepository<Organization, Long> {
