+++ src/main/java/KUSITMS/WITHUS/domain/organization/organization/service/OrganizationService.java
@@ -1,7 +1,7 @@
-package KUSITMS.WITHUS.domain.organization.service;
+package KUSITMS.WITHUS.domain.organization.organization.service;
 
-import KUSITMS.WITHUS.domain.organization.dto.OrganizationRequestDTO;
-import KUSITMS.WITHUS.domain.organization.dto.OrganizationResponseDTO;
+import KUSITMS.WITHUS.domain.organization.organization.dto.OrganizationRequestDTO;
+import KUSITMS.WITHUS.domain.organization.organization.dto.OrganizationResponseDTO;
 
 import java.util.List;
