+++ src/main/java/KUSITMS/WITHUS/domain/organization/organizationRole/entity/OrganizationRole.java
@@ -1,0 +1,4 @@
+package KUSITMS.WITHUS.domain.organization.organizationRole.entity;
+
+public class OrganizationRole {
+}
