+++ src/main/java/KUSITMS/WITHUS/domain/recruitment/entity/Recruitment.java
@@ -2,7 +2,7 @@ package KUSITMS.WITHUS.domain.recruitment.entity;
 
 import KUSITMS.WITHUS.domain.application.template.entity.ApplicationTemplate;
 import KUSITMS.WITHUS.domain.evaluation.evaluationCriteria.entity.EvaluationCriteria;
-import KUSITMS.WITHUS.domain.organization.entity.Organization;
+import KUSITMS.WITHUS.domain.organization.organization.entity.Organization;
 import KUSITMS.WITHUS.global.common.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
